create separate page for user dashboard and remove component dashboard created
