--- src/webapp/src/pages/DashboardPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import UserDashboard from '../components/dashboard/UserDashboard';
-
-const Dashboard = () => {
-  return <div>
-    <UserDashboard />
-  </div>;
-};
-
-export default Dashboard;
