+++ src/webapp/src/pages/index.js
@@ -1,4 +1,4 @@
 export { default as HomePage } from './HomePage';
 export { default as AboutPage } from './AboutPage';
 export { default as NotFoundPage } from './NotFoundPage';
-export { default as Dashboard } from './DashboardPage';
+export { default as UserDashboard } from './UserDashboardPage';
